Change :  take id from params not body

+++ part4/controllers/blogs.js
@@ -19,13 +19,29 @@ blogRouter.get("/", async (request, response, next) => {
 
 blogRouter.delete("/:id", async (request, response, next) => {
   try {
-    const id = request.body.id;
-    const deletedBlog = await Blog.findByIdAndDelete(id);
+    const id = request.params.id;
+    const token = request.token;
+    console.log(token, "token from middleware");
+    if (!token) {
+      return response.status(401).json({ error: "token missing" });
+    }
+    const decodedToken = jwt.verify(token, config.SEKRET);
 
-    if (!deletedBlog) {
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "token invalid" });
+    }
+
+    const blog = await Blog.findById(id);
+    if (!blog) {
       return response.status(404).json({ error: "blog not found" });
     }
 
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+      return response.status(401).json({
+        error: "only the creator can delete this blog",
+      });
+    }
+    await Blog.findByIdAndDelete(id);
     response.status(204).end();
   } catch (error) {
     next(error);
@@ -64,20 +80,18 @@ blogRouter.put("/:id", async (request, response, next) => {
 
 blogRouter.post("/", async (request, response, next) => {
   try {
-
     // const authorization = request.get("authorization");
     // console.log(authorization, "authorization");
     // const authoArr = authorization && authorization.split(" ");
-    
+
     const decodedToken = jwt.verify(request.token, config.SEKRET);
 
     console.log(decodedToken, "this is decoded token");
 
     console.log(request.token, "this is token fron request.token");
 
-
-    if(!decodedToken){
-      return response.status(401).json({error:"Unauthorized"});
+    if (!decodedToken) {
+      return response.status(401).json({ error: "Unauthorized" });
     }
 
     const user = await User.findById(decodedToken.id);
@@ -88,8 +102,7 @@ blogRouter.post("/", async (request, response, next) => {
         .status(400)
         .json({ error: "userId missing or not valid" });
     }
-    const { title, author, url, likes} = request.body;
-
+    const { title, author, url, likes } = request.body;
 
     const blog = new Blog({
       title: title,
